+++ NEWS.md
@@ -1,3 +1,10 @@
+# cvasi 1.1.0
+
+* `calibrate()`
+  * Function arguments simplified and partly renamed
+  * Fitting is now more error-tolerant and will not abort if a simulation
+    fails or if it returns invalid values
+
 # cvasi 1.0.1
 
 * `Algae_TKTD()` harmonized with `Algae_Simple()` model equations